자료형 표기 통일

+++ contracts/sale/Sale.sol
@@ -192,7 +192,7 @@ contract Sale is Stateable {
         }
     }
 
-    function buyerAddressTransfer(uint _id, address _product, address _from, address _to)
+    function buyerAddressTransfer(uint256 _id, address _product, address _from, address _to)
         external
         onlyOwner
         validAddress(_product)
